+++ NEWS.md
@@ -4,6 +4,8 @@ Pentobi Release Notes
 Version 26.0-dev (main)
 -----------------------
 
+* Initial French translation of the manual (still rough and in need
+  of revision by a native speaker)
 * The minimum required Qt version is now 6.5.
 * Removed dependency on qml6-module-qt-labs-settings (on Debian).
 * itstool is no longer required during the normal build process,